pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "fastapi>=0.121.3",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "langchain-groq>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "gunicorn>=21.2.0",
]
